refactor(renderer/api/ai): move interfaces to co-located types/ folder
Move exported interfaces from ai module files to api/ai/types/:
- MihtnelisAgentStreamEvent, MihtnelisAgentStreamRequest
- ResolveMihtnelisWebAccessRequest, ResolveMihtnelisLocalToolAccessRequest,
  ResolveMihtnelisLocalToolRequest
- FetchAgentPromptRequest, FetchAgentPromptResponse
- CustomDirectAgentRequest, OllamaLocalAgentRequest
- RealtimeSttEvent, RealtimeSttSession, StartRealtimeSttRequest
