+++ src/renderer/api/ai/customDirectAgent.ts
@@ -25,7 +25,10 @@
  */
 
 import { fetchAgentPrompt } from './mihtnelisAgentStream';
-import type { MihtnelisAgentStreamEvent, MihtnelisAgentStreamEventType } from './mihtnelisAgentStream';
+import type { MihtnelisAgentStreamEvent } from './types/MihtnelisAgentStreamEvent';
+import type { CustomDirectAgentRequest } from './types/CustomDirectAgentRequest';
+
+export type { CustomDirectAgentRequest };
 
 let cachedSystemPrompt: string | null = null;
 let cachedPromptKey = '';
@@ -71,22 +74,6 @@ async function resolveSystemPrompt(request: CustomDirectAgentRequest): Promise<s
 export function clearCustomDirectPromptCache(): void {
   cachedSystemPrompt = null;
   cachedPromptKey = '';
-}
-
-export interface CustomDirectAgentRequest {
-  token: string;
-  message: string;
-  model: string;
-  agentMode?: string;
-  context?: string;
-  workspaces?: string[];
-  skills?: Array<{ name: string; content: string }>;
-  snapshotMode?: boolean;
-  baseUrl: string;
-  apiKey: string;
-  temperature?: number;
-  signal?: AbortSignal;
-  onEvent?: (event: MihtnelisAgentStreamEvent) => void;
 }
 
 /**
