+++ src/renderer/api/ai/mihtnelisAgentStream.ts
@@ -20,71 +20,27 @@
  */
 
 import { buildReplayHeaders, resolveClientVersion, USER_ACCOUNT_API_BASE } from '../user/userAccountApi.client';
+import type { MihtnelisAgentStreamEvent, MihtnelisAgentStreamEventType } from './types/MihtnelisAgentStreamEvent';
+import type { ResolveMihtnelisWebAccessRequest } from './types/ResolveMihtnelisWebAccessRequest';
+import type { ResolveMihtnelisLocalToolAccessRequest } from './types/ResolveMihtnelisLocalToolAccessRequest';
+import type { ResolveMihtnelisLocalToolRequest } from './types/ResolveMihtnelisLocalToolRequest';
+import type { MihtnelisAgentStreamRequest } from './types/MihtnelisAgentStreamRequest';
+import type { FetchAgentPromptRequest } from './types/FetchAgentPromptRequest';
+import type { FetchAgentPromptResponse } from './types/FetchAgentPromptResponse';
+
+export type {
+  MihtnelisAgentStreamEvent,
+  MihtnelisAgentStreamEventType,
+  ResolveMihtnelisWebAccessRequest,
+  ResolveMihtnelisLocalToolAccessRequest,
+  ResolveMihtnelisLocalToolRequest,
+  MihtnelisAgentStreamRequest,
+  FetchAgentPromptRequest,
+  FetchAgentPromptResponse,
+};
 
 const APP_NAME_HEADER = 'X-App-Name';
 const APP_NAME_VALUE = 'eisland';
-
-export type MihtnelisAgentStreamEventType =
-  | 'meta'
-  | 'tool'
-  | 'tool_call_request'
-  | 'tool_call_result'
-  | 'think'
-  | 'chunk'
-  | 'chunk_reset'
-  | 'billing'
-  | 'web_access_request'
-  | 'web_access_resolved'
-  | 'todo'
-  | 'final'
-  | 'error';
-
-export interface MihtnelisAgentStreamEvent {
-  type: MihtnelisAgentStreamEventType;
-  payload: unknown;
-}
-
-export interface ResolveMihtnelisWebAccessRequest {
-  token: string;
-  requestId: string;
-  allow: boolean;
-}
-
-export interface ResolveMihtnelisLocalToolAccessRequest {
-  token: string;
-  requestId: string;
-  allow: boolean;
-}
-
-export interface ResolveMihtnelisLocalToolRequest {
-  token: string;
-  requestId: string;
-  success: boolean;
-  result?: unknown;
-  error?: string;
-  durationMs?: number;
-}
-
-export interface MihtnelisAgentStreamRequest {
-  token: string;
-  message: string;
-  sessionId?: string;
-  provider?: string;
-  model?: string;
-  agentMode?: string;
-  context?: string;
-  workspaces?: string[];
-  skills?: Array<{ name: string; content: string }>;
-  thinking?: boolean;
-  reasoningEffort?: 'low' | 'medium' | 'high';
-  timestamp?: string;
-  location?: string;
-  snapshotMode?: boolean;
-  customApiKey?: string;
-  customEndpoint?: string;
-  signal?: AbortSignal;
-  onEvent?: (event: MihtnelisAgentStreamEvent) => void;
-}
 
 /**
  * 发起 mihtnelis agent 流式请求。
@@ -318,21 +274,6 @@ export async function resolveMihtnelisLocalToolAccess(request: ResolveMihtnelisL
     const body = await response.text().catch(() => '');
     throw new Error(`本地工具授权提交失败 (${response.status}): ${body || response.statusText}`);
   }
-}
-
-export interface FetchAgentPromptRequest {
-  token: string;
-  agentMode?: string;
-  snapshotMode?: boolean;
-  localMode?: boolean;
-  workspaces?: string[];
-  skills?: Array<{ name: string; content: string }>;
-}
-
-export interface FetchAgentPromptResponse {
-  success: boolean;
-  systemPrompt: string;
-  error?: string;
 }
 
 /**
