+++ src/renderer/api/ai/types/FetchAgentPromptResponse.ts
@@ -1,0 +1,35 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file FetchAgentPromptResponse.ts
+ * @description 获取 Agent 提示词响应类型定义
+ * @author 鸡哥
+ */
+
+/** 获取 Agent 提示词响应 */
+export interface FetchAgentPromptResponse {
+  /** 请求是否成功 */
+  success: boolean;
+  /** 系统提示词 */
+  systemPrompt: string;
+  /** 错误信息 */
+  error?: string;
+}
