+++ src/renderer/api/ai/types/RealtimeSttEvent.ts
@@ -1,0 +1,36 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file RealtimeSttEvent.ts
+ * @description 实时语音识别事件类型定义
+ * @author 鸡哥
+ */
+
+/** 实时语音识别事件类型 */
+export type RealtimeSttEventType = 'partial' | 'final' | 'error' | 'ready';
+
+/** 实时语音识别事件 */
+export interface RealtimeSttEvent {
+  /** 事件类型 */
+  type: RealtimeSttEventType;
+  /** 识别文本 */
+  text: string;
+}
