+++ src/renderer/api/ai/types/ResolveMihtnelisLocalToolAccessRequest.ts
@@ -1,0 +1,35 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file ResolveMihtnelisLocalToolAccessRequest.ts
+ * @description 解决 mihtnelis 本地工具访问请求类型定义
+ * @author 鸡哥
+ */
+
+/** 解决 mihtnelis 本地工具访问请求 */
+export interface ResolveMihtnelisLocalToolAccessRequest {
+  /** 用户 token */
+  token: string;
+  /** 请求 ID */
+  requestId: string;
+  /** 是否允许访问 */
+  allow: boolean;
+}
