+++ src/renderer/api/ai/types/index.ts
@@ -1,0 +1,38 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file index.ts
+ * @description AI 模块类型定义导出
+ * @author 鸡哥
+ */
+
+export type { MihtnelisAgentStreamEvent, MihtnelisAgentStreamEventType } from './MihtnelisAgentStreamEvent';
+export type { ResolveMihtnelisWebAccessRequest } from './ResolveMihtnelisWebAccessRequest';
+export type { ResolveMihtnelisLocalToolAccessRequest } from './ResolveMihtnelisLocalToolAccessRequest';
+export type { ResolveMihtnelisLocalToolRequest } from './ResolveMihtnelisLocalToolRequest';
+export type { MihtnelisAgentStreamRequest } from './MihtnelisAgentStreamRequest';
+export type { FetchAgentPromptRequest } from './FetchAgentPromptRequest';
+export type { FetchAgentPromptResponse } from './FetchAgentPromptResponse';
+export type { CustomDirectAgentRequest } from './CustomDirectAgentRequest';
+export type { OllamaLocalAgentRequest } from './OllamaLocalAgentRequest';
+export type { RealtimeSttEvent, RealtimeSttEventType } from './RealtimeSttEvent';
+export type { RealtimeSttSession } from './RealtimeSttSession';
+export type { StartRealtimeSttRequest } from './StartRealtimeSttRequest';
